+++ modules/core/src/main/java/net/ashald/envfile/EnvVarsProvider.java
@@ -1,17 +1,20 @@
 package net.ashald.envfile;
 
-import org.jetbrains.annotations.NotNull;
+import net.ashald.envfile.exceptions.EnvFileException;
 
 import java.io.File;
-import java.io.IOException;
 import java.util.Map;
-
 
 public interface EnvVarsProvider {
 
-    @NotNull Map<String, String> process(@NotNull Map<String, String> runConfigEnv, String path, @NotNull Map<String, String> aggregatedEnv) throws EnvFileErrorException, IOException;
+    /**
+     * @param file         file to process
+     * @param isExecutable if true, file should be executed and its output should be parsed
+     * @param context      environment variables that were generated so far by other providers
+     * @return environment variables defined by files, does not need to merge with context
+     * @throws EnvFileException if something goes wrong
+     */
+    Map<String, String> getEnvVars(File file, boolean isExecutable, Map<String, String> context)
+            throws EnvFileException;
 
-    boolean isEditable();
-
-    boolean isFileLocationValid(File file);
 }