+++ modules/platform/src/main/java/net/ashald/envfile/platform/ui/EnvFileConfigurationPanel.java
@@ -10,37 +10,47 @@ import com.intellij.openapi.fileChooser.FileChooser;
 import com.intellij.openapi.fileChooser.FileChooserDescriptor;
 import com.intellij.openapi.fileChooser.FileChooserDescriptorFactory;
 import com.intellij.openapi.project.Project;
-import com.intellij.openapi.ui.popup.JBPopupFactory;
 import com.intellij.openapi.ui.popup.ListPopup;
 import com.intellij.openapi.util.Conditions;
 import com.intellij.openapi.vfs.VirtualFile;
-import com.intellij.ui.*;
+import com.intellij.ui.AnActionButton;
+import com.intellij.ui.AnActionButtonUpdater;
+import com.intellij.ui.BooleanTableCellRenderer;
+import com.intellij.ui.ToolbarDecorator;
 import com.intellij.ui.popup.PopupFactoryImpl;
 import com.intellij.ui.table.TableView;
 import com.intellij.util.ui.ColumnInfo;
 import com.intellij.util.ui.JBUI;
 import com.intellij.util.ui.ListTableModel;
+import lombok.val;
+import net.ashald.envfile.EnvVarsProviderFactory;
 import net.ashald.envfile.platform.EnvFileEntry;
-import net.ashald.envfile.platform.EnvVarsProviderExtension;
 import net.ashald.envfile.platform.EnvFileSettings;
+import net.ashald.envfile.platform.EnvVarsProviderExtension;
 import net.ashald.envfile.platform.ui.table.EnvFileIsActiveColumnInfo;
+import net.ashald.envfile.platform.ui.table.EnvFileIsExecutableColumnInfo;
 import net.ashald.envfile.platform.ui.table.EnvFilePathColumnInfo;
 import net.ashald.envfile.platform.ui.table.EnvFileTypeColumnInfo;
+import net.ashald.envfile.providers.runconfig.RunConfigEnvVarsProvider;
 
-import javax.swing.*;
-import javax.swing.border.EmptyBorder;
+import javax.swing.BoxLayout;
+import javax.swing.JCheckBox;
+import javax.swing.JPanel;
 import javax.swing.table.JTableHeader;
 import javax.swing.table.TableColumn;
-import java.awt.*;
+import java.awt.BorderLayout;
+import java.awt.Dimension;
+import java.awt.FontMetrics;
 import java.awt.event.ActionEvent;
 import java.awt.event.ActionListener;
-import java.util.*;
+import java.util.ArrayList;
+import java.util.LinkedList;
 
 
-class EnvFileConfigurationPanel<T extends RunConfigurationBase> extends JPanel {
+class EnvFileConfigurationPanel<T extends RunConfigurationBase<?>> extends JPanel {
     private static final int MAX_RECENT = 5;
-    private static final LinkedList<EnvFileEntry> recent = new LinkedList<EnvFileEntry>();
-    private final RunConfigurationBase runConfig;
+    private static final LinkedList<EnvFileEntry> RECENT = new LinkedList<>();
+    private final RunConfigurationBase<?> runConfig;
 
     private final JCheckBox useEnvFileCheckBox;
     private final JCheckBox substituteEnvVarsCheckBox;
@@ -54,18 +64,20 @@ class EnvFileConfigurationPanel<T extends RunConfigurationBase> extends JPanel {
         runConfig = configuration;
 
         // Define Model
-        ColumnInfo<EnvFileEntry, Boolean> IS_ACTIVE = new EnvFileIsActiveColumnInfo();
-        ColumnInfo<EnvFileEntry, String> FILE = new EnvFilePathColumnInfo();
-        ColumnInfo<EnvFileEntry, EnvFileEntry> TYPE = new EnvFileTypeColumnInfo();
+        ColumnInfo<EnvFileEntry, Boolean> columnIsActive = new EnvFileIsActiveColumnInfo();
+        ColumnInfo<EnvFileEntry, Boolean> columnIsExecutable = new EnvFileIsExecutableColumnInfo();
+        ColumnInfo<EnvFileEntry, String> columnFile = new EnvFilePathColumnInfo(runConfig.getProject());
+        ColumnInfo<EnvFileEntry, EnvFileEntry> columnType = new EnvFileTypeColumnInfo();
 
-        envFilesModel = new ListTableModel<>(IS_ACTIVE, FILE, TYPE);
+        envFilesModel = new ListTableModel<>(columnIsActive, columnIsExecutable, columnFile, columnType);
 
         // Create Table
         envFilesTable = new TableView<>(envFilesModel);
         envFilesTable.getEmptyText().setText("No environment variables files selected");
 
-        setUpColumnWidth(envFilesTable, 0, IS_ACTIVE, 20);
-        setUpColumnWidth(envFilesTable, 2, TYPE, 50);
+        setUpColumnWidth(envFilesTable, 0, columnIsActive, 20);
+        setUpColumnWidth(envFilesTable, 1, columnIsExecutable, 20);
+        setUpColumnWidth(envFilesTable, 3, columnType, 50);
 
         envFilesTable.setColumnSelectionAllowed(false);
         envFilesTable.setShowGrid(false);
@@ -83,49 +95,47 @@ class EnvFileConfigurationPanel<T extends RunConfigurationBase> extends JPanel {
                 substituteEnvVarsCheckBox.setEnabled(useEnvFileCheckBox.isSelected());
                 supportPathMacroCheckBox.setEnabled(useEnvFileCheckBox.isSelected());
                 ignoreMissingCheckBox.setEnabled(useEnvFileCheckBox.isSelected());
+                experimentalIntegrationsCheckBox.setEnabled(useEnvFileCheckBox.isSelected());
             }
         });
         substituteEnvVarsCheckBox = new JCheckBox("Substitute Environment Variables (${FOO} / ${BAR:-default} / $${ESCAPED})");
-        substituteEnvVarsCheckBox.addActionListener(e -> envFilesModel.getItems().forEach(envFileEntry -> envFileEntry.setSubstitutionEnabled(substituteEnvVarsCheckBox.isSelected())));
         supportPathMacroCheckBox = new JCheckBox("Process JetBrains path macro references ($PROJECT_DIR$)");
         ignoreMissingCheckBox = new JCheckBox("Ignore missing files");
         experimentalIntegrationsCheckBox = new JCheckBox("Enable experimental integrations (e.g. Gradle) - may break any time!");
 
         // TODO: come up with a generic approach for this
-        envFilesModel.addRow(new EnvFileEntry(runConfig, "runconfig", null, true, substituteEnvVarsCheckBox.isSelected()));
+        envFilesModel.addRow(
+                EnvFileEntry.builder()
+                        .parserId(RunConfigEnvVarsProvider.PARSER_ID)
+                        .enabled(true)
+                        .executable(false)
+                        .build()
+        );
 
         // Create Toolbar - Add/Remove/Move actions
         final ToolbarDecorator envFilesTableDecorator = ToolbarDecorator.createDecorator(envFilesTable);
 
-        final AnActionButtonUpdater updater = new AnActionButtonUpdater() {
-            @Override
-            public boolean isEnabled(AnActionEvent e) {
-                return useEnvFileCheckBox.isSelected();
-            }
-        };
+        final AnActionButtonUpdater updater = e -> useEnvFileCheckBox.isSelected();
 
         envFilesTableDecorator
-                .setAddAction(new AnActionButtonRunnable() {
-            @Override
-            public void run(AnActionButton button) {
-                doAddAction(button, envFilesTable, envFilesModel);
-            }
-        })
+                .setAddAction(button -> doAddAction(button, envFilesTable, envFilesModel))
                 .setAddActionUpdater(updater)
-                .setRemoveActionUpdater(new AnActionButtonUpdater() {
-                    @Override
-                    public boolean isEnabled(AnActionEvent e) {
-                        boolean allEditable = true;
+                .setRemoveActionUpdater(e -> {
+                    boolean allEditable = true;
 
-                        for (EnvFileEntry entry : envFilesTable.getSelectedObjects()) {
-                            if (!entry.isEditable()) {
-                                allEditable = false;
-                                break;
-                            }
+                    for (EnvFileEntry entry : envFilesTable.getSelectedObjects()) {
+                        val editable =
+                                EnvVarsProviderExtension.getParserFactoryById(entry.getParserId())
+                                        .map(EnvVarsProviderFactory::isEditable)
+                                        .orElse(true);
+
+                        if (!editable) {
+                            allEditable = false;
+                            break;
                         }
-
-                        return updater.isEnabled(e) && envFilesTable.getSelectedRowCount() >= 1 && allEditable;
                     }
+
+                    return updater.isEnabled(e) && envFilesTable.getSelectedRowCount() >= 1 && allEditable;
                 });
 
         JPanel optionsPanel = new JPanel();
@@ -155,7 +165,7 @@ class EnvFileConfigurationPanel<T extends RunConfigurationBase> extends JPanel {
         add(envFilesTableDecoratorPanel, BorderLayout.CENTER);
     }
 
-    private void setUpColumnWidth(TableView<EnvFileEntry> table, int columnIdx, ColumnInfo columnInfo, int extend) {
+    private void setUpColumnWidth(TableView<EnvFileEntry> table, int columnIdx, ColumnInfo<?, ?> columnInfo, int extend) {
         JTableHeader tableHeader = table.getTableHeader();
         FontMetrics fontMetrics = tableHeader.getFontMetrics(tableHeader.getFont());
 
@@ -169,11 +179,15 @@ class EnvFileConfigurationPanel<T extends RunConfigurationBase> extends JPanel {
         tableColumn.setMaxWidth(preferredWidth);
     }
 
-    private void doAddAction(AnActionButton button, final TableView<EnvFileEntry> table, final ListTableModel<EnvFileEntry> model) {
-        DefaultActionGroup actionGroup = new DefaultActionGroup(null, false);
+    private void doAddAction(
+            AnActionButton button,
+            final TableView<EnvFileEntry> table,
+            final ListTableModel<EnvFileEntry> model
+    ) {
+        DefaultActionGroup actionGroup = new DefaultActionGroup();
 
         for (final EnvVarsProviderExtension extension : EnvVarsProviderExtension.getParserExtensions()) {
-            if (!extension.getFactory().createProvider(substituteEnvVarsCheckBox.isSelected()).isEditable()) {
+            if (!extension.getFactory().isEditable()) {
                 continue;
             }
 
@@ -197,29 +211,39 @@ class EnvFileConfigurationPanel<T extends RunConfigurationBase> extends JPanel {
                         }
 
                         ArrayList<EnvFileEntry> newList = new ArrayList<EnvFileEntry>(model.getItems());
-                        final EnvFileEntry newOptions = new EnvFileEntry(runConfig, extension.getId(), selectedPath, true, substituteEnvVarsCheckBox.isSelected());
+                        final EnvFileEntry newOptions = EnvFileEntry.builder()
+                                .parserId(extension.getId())
+                                .path(selectedPath)
+                                .enabled(true)
+                                .executable(false)
+                                .build();
+
                         newList.add(newOptions);
                         model.setItems(newList);
                         int index = model.getRowCount() - 1;
                         model.fireTableRowsInserted(index, index);
                         table.setRowSelectionInterval(index, index);
 
-                        synchronized (recent) {
-                            recent.remove(newOptions);
-                            recent.addFirst(newOptions);
-                            if (recent.size() > MAX_RECENT) recent.removeLast();
+                        synchronized (RECENT) {
+                            RECENT.remove(newOptions);
+                            RECENT.addFirst(newOptions);
+                            if (RECENT.size() > MAX_RECENT) RECENT.removeLast();
                         }
                     }
                 }
             };
             actionGroup.add(anAction);
         }
-        synchronized (recent) {
-            if (!recent.isEmpty()) {
+        synchronized (RECENT) {
+            if (!RECENT.isEmpty()) {
                 actionGroup.addSeparator("Recent");
 
-                for (final EnvFileEntry entry : recent) {
-                    String title = String.format("%s -> %s", entry.getTypeTitle(), entry.getPath());
+                for (final EnvFileEntry entry : RECENT) {
+                    val typeTitle = EnvVarsProviderExtension.getParserFactoryById(entry.getParserId())
+                            .map(EnvVarsProviderFactory::getTitle)
+                            .orElse(String.format("<%s>", entry.getParserId()));
+
+                    String title = String.format("%s -> %s", typeTitle, entry.getPath());
                     String shortTitle = title.length() < 81 ? title : title.replaceFirst("(.{39}).+(.{39})", "$1...$2");
                     AnAction anAction = new AnAction(shortTitle, title, null) {
                         @Override
@@ -230,9 +254,9 @@ class EnvFileConfigurationPanel<T extends RunConfigurationBase> extends JPanel {
                             int index = model.getRowCount() - 1;
                             model.fireTableRowsInserted(index, index);
                             table.setRowSelectionInterval(index, index);
-                            synchronized (recent) {
-                                recent.remove(entry);
-                                recent.addFirst(entry);
+                            synchronized (RECENT) {
+                                RECENT.remove(entry);
+                                RECENT.addFirst(entry);
                             }
                         }
                     };
@@ -248,33 +272,34 @@ class EnvFileConfigurationPanel<T extends RunConfigurationBase> extends JPanel {
                         false, false, false, false,
                         null, -1, Conditions.<AnAction>alwaysTrue(), popupPlace);
 
-        popup.show(button.getPreferredPopupPoint());
-
+        if (button.getPreferredPopupPoint() != null) {
+            popup.show(button.getPreferredPopupPoint());
+        }
     }
 
     EnvFileSettings getState() {
-        return new EnvFileSettings(
-                useEnvFileCheckBox.isSelected(),
-                substituteEnvVarsCheckBox.isSelected(),
-                supportPathMacroCheckBox.isSelected(),
-                envFilesModel.getItems(),
-                ignoreMissingCheckBox.isSelected(),
-                experimentalIntegrationsCheckBox.isSelected()
-        );
+        return EnvFileSettings.builder()
+                .pluginEnabled(useEnvFileCheckBox.isSelected())
+                .envVarsSubstitutionEnabled(substituteEnvVarsCheckBox.isSelected())
+                .pathMacroSupported(supportPathMacroCheckBox.isSelected())
+                .ignoreMissing(ignoreMissingCheckBox.isSelected())
+                .enableExperimentalIntegrations(experimentalIntegrationsCheckBox.isSelected())
+                .entries(envFilesModel.getItems())
+                .build();
     }
 
     void setState(EnvFileSettings state) {
-        useEnvFileCheckBox.setSelected(state.isEnabled());
+        useEnvFileCheckBox.setSelected(state.isPluginEnabledEnabled());
         substituteEnvVarsCheckBox.setSelected(state.isSubstituteEnvVarsEnabled());
         supportPathMacroCheckBox.setSelected(state.isPathMacroSupported());
         ignoreMissingCheckBox.setSelected(state.isIgnoreMissing());
         experimentalIntegrationsCheckBox.setSelected(state.isEnableExperimentalIntegrations());
 
-        envFilesTable.setEnabled(state.isEnabled());
-        substituteEnvVarsCheckBox.setEnabled(state.isEnabled());
-        supportPathMacroCheckBox.setEnabled(state.isEnabled());
-        ignoreMissingCheckBox.setEnabled(state.isEnabled());
-        experimentalIntegrationsCheckBox.setEnabled(state.isEnabled());
+        envFilesTable.setEnabled(state.isPluginEnabledEnabled());
+        substituteEnvVarsCheckBox.setEnabled(state.isPluginEnabledEnabled());
+        supportPathMacroCheckBox.setEnabled(state.isPluginEnabledEnabled());
+        ignoreMissingCheckBox.setEnabled(state.isPluginEnabledEnabled());
+        experimentalIntegrationsCheckBox.setEnabled(state.isPluginEnabledEnabled());
 
         envFilesModel.setItems(new ArrayList<>(state.getEntries()));
     }