+++ modules/products/idea/src/main/java/net/ashald/envfile/products/idea/scala/SbtRunEnvFileConfigurationForm.java
@@ -2,6 +2,7 @@ package net.ashald.envfile.products.idea.scala;
 
 import com.intellij.execution.ExecutionException;
 import com.intellij.openapi.project.Project;
+import net.ashald.envfile.platform.EnvFileEnvironmentVariables;
 import net.ashald.envfile.platform.ui.EnvFileConfigurationEditor;
 import org.jetbrains.sbt.runner.SbtRunConfiguration;
 import org.jetbrains.sbt.runner.SbtRunConfigurationForm;
@@ -9,7 +10,7 @@ import org.jetbrains.sbt.runner.SbtRunConfigurationForm;
 import java.util.Map;
 
 public class SbtRunEnvFileConfigurationForm extends SbtRunConfigurationForm {
-    private SbtRunConfiguration configuration;
+    private final SbtRunConfiguration configuration;
 
     public SbtRunEnvFileConfigurationForm(Project project, SbtRunConfiguration sbtRunConfiguration) {
         super(project, sbtRunConfiguration);
@@ -18,11 +19,23 @@ public class SbtRunEnvFileConfigurationForm extends SbtRunConfigurationForm {
 
     @Override
     public Map<String, String> getEnvironmentVariables() {
+        Map<String, String> baseEnv = super.getEnvironmentVariables();
+        Map<String, String> newEnv;
         try {
-            return EnvFileConfigurationEditor.collectEnv(configuration, super.getEnvironmentVariables());
+            newEnv = new EnvFileEnvironmentVariables(
+                    EnvFileConfigurationEditor.getEnvFileSetting(configuration)
+            )
+                    .render(
+                            configuration.getProject(),
+                            baseEnv,
+                            true // can we get it dynamically?
+                    );
         } catch (ExecutionException e) {
-            e.printStackTrace();
+            throw new RuntimeException(e);
         }
-        return super.getEnvironmentVariables();
+
+        return newEnv == null
+                ? super.getEnvironmentVariables()
+                : newEnv;
     }
 }