+++ modules/products/rubymine/src/main/java/net/ashald/envfile/products/rubymine/RubyMineRunConfigurationExtension.java
@@ -6,6 +6,7 @@ import com.intellij.execution.configurations.RunnerSettings;
 import com.intellij.openapi.options.SettingsEditor;
 import com.intellij.openapi.util.InvalidDataException;
 import com.intellij.openapi.util.WriteExternalException;
+import net.ashald.envfile.platform.EnvFileEnvironmentVariables;
 import net.ashald.envfile.platform.ui.EnvFileConfigurationEditor;
 import org.jdom.Element;
 import org.jetbrains.annotations.NotNull;
@@ -13,7 +14,6 @@ import org.jetbrains.annotations.Nullable;
 import org.jetbrains.plugins.ruby.ruby.run.configuration.AbstractRubyRunConfiguration;
 import org.jetbrains.plugins.ruby.ruby.run.configuration.RubyRunConfigurationExtension;
 
-import java.util.HashMap;
 import java.util.Map;
 
 public class RubyMineRunConfigurationExtension extends RubyRunConfigurationExtension {
@@ -52,11 +52,27 @@ public class RubyMineRunConfigurationExtension extends RubyRunConfigurationExten
     }
 
     @Override
-    protected void patchCommandLine(@NotNull AbstractRubyRunConfiguration<?> configuration, @Nullable RunnerSettings runnerSettings, @NotNull GeneralCommandLine cmdLine, @NotNull String runnerId) throws ExecutionException {
-        Map<String, String> currentEnv = cmdLine.getEnvironment();
-        Map<String, String> newEnv = EnvFileConfigurationEditor.collectEnv(configuration, new HashMap<>(currentEnv));
-        currentEnv.clear();
-        currentEnv.putAll(newEnv);
+    protected void patchCommandLine(
+            @NotNull AbstractRubyRunConfiguration<?> configuration,
+            @Nullable RunnerSettings runnerSettings,
+            @NotNull GeneralCommandLine cmdLine,
+            @NotNull String runnerId
+    ) throws ExecutionException {
+        Map<String, String> newEnv = new EnvFileEnvironmentVariables(
+                EnvFileConfigurationEditor.getEnvFileSetting(configuration)
+        )
+                .render(
+                        configuration.getProject(),
+                        cmdLine.getEnvironment(),
+                        cmdLine.isPassParentEnvironment()
+                );
+
+        if (newEnv == null) {
+            return;
+        }
+
+        cmdLine.getEnvironment().clear();
+        cmdLine.getEnvironment().putAll(newEnv);
     }
 
     //