+++ CHANGELOG.md
@@ -2,6 +2,39 @@
 All notable changes to this project will be documented in this file.
 This project adheres to [Semantic Versioning].
 
+## 3.4.2 - 2023-12-25
+
+### Fixed
+
+- Bump `org.apache.commons:commons-text` to `1.10.0'
+
+## 3.4.1 - 2022-10-04
+
+### Fixed
+
+- Plugin affecting env vars even when disabled
+
+## 3.4.0 - 2022-10-03
+
+### Added
+
+- When environment variables file is executed, it now receives all environment variables accumulated by previous entries
+- Log stderr from execution of environment variables files into IDE event console
+
+## 3.3.0 - 2022-10-02
+
+### Added
+
+- Multi-line variable support in `.env` files ([#127]) - special thanks to [@ledoyen](https://github.com/ledoyen)
+- Ignore for `export ` prefix in `.env` files
+- Include parent environment variables (when enabled) into environment substitution ([#165]) - special thanks to [@jansorg](https://github.com/jansorg)
+- An option to execute file and parse content from standard output
+
+### Fixed
+
+- Incompatibility with IDEA 2022.2 ([#151]) - special thanks to [@HassanAbouelela](https://github.com/HassanAbouelela)
+- Experimental features checkbox is properly enabled/disabled when plugin is enabled/disabled
+
 ## 3.2.2 - 2021-09-02
 
 ### Fixed
@@ -126,7 +159,9 @@ This project adheres to [Semantic Versioning].
 [#81]: https://github.com/Ashald/EnvFile/issues/81
 [#81]: https://github.com/Ashald/EnvFile/issues/83
 [#100]: https://github.com/ashald/EnvFile/issues/100
+[#127]: https://github.com/ashald/EnvFile/issues/127
 [#151]: https://github.com/ashald/EnvFile/issues/151
+[#165]: https://github.com/ashald/EnvFile/issues/165
 
 [Keep a CHANGELOG]:     http://keepachangelog.com
 [Semantic Versioning]:  http://semver.org/